pyproject.toml
[tool.poetry]
name = "bng36-wgs84-conversion"
version = "0.1.0"
description = "Python Poetry Boilerplate"
authors = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
maintainers = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/face0b1101/bng36-wgs84-conversion"
repository = "https://github.com/face0b1101/bng36-wgs84-conversion"
documentation = "https://github.com/face0b1101/bng36-wgs84-conversion"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.6"
Unipath = "^1.1"
rich = "^12.5.1"
pyproj = "^3.6.1"
fastapi = "^0.109.0"
pydantic = "^2.5.3"
uvicorn = {extras = ["standard"], version = "^0.27.0"}

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
ipython = "^8.0.0"
jupyter = "^1.0.0"
pytest = "^6.2.5"
tox = "^3.25.1"
isort = "^5.10.1"
scrapy-gui = "^1.2"
flake8-bugbear = "^23.9.16"
httpx = "^0.26.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
# make it compatible with black
profile = "black"

[tool.poetry.scripts]
bng36-wgs84-conversion = "bng36_wgs84_conversion.main:main"
